% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOCanalyses-package.R
\docType{package}
\name{SOCanalyses-package}
\alias{SOCanalyses}
\alias{SOCanalyses-package}
\title{SOCanalyses: The Title of the Project}
\description{
This package analyses the results of meta-analyses studying the effect of human interventions on soil carbon. The analyses are based on a systematic synthesis of the available literature on the CIRAD repository (https://doi.org/10.18167/DVN1/KKPLR834). The database is described in Beillouin et al. 2022 (https://doi.org/10.1038/s41597-022-01318-1).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dbeillouin/SOCanalyses}
  \item Report bugs at \url{https://github.com/dbeillouin/SOCanalyses/issues}
}

}
\author{
\strong{Maintainer}: Damien Beillouin \email{d.beillouin@cirad.fr} (\href{https://orcid.org/{{0000-0003-2014-3482}}}{ORCID}) [copyright holder]

}
\keyword{internal}
