% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOCanalyses-package.R
\docType{package}
\name{SOCanalyses-package}
\alias{SOCanalyses}
\alias{SOCanalyses-package}
\title{SOCanalyses: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dbeillouin/SOCanalyses}
  \item Report bugs at \url{https://github.com/dbeillouin/SOCanalyses/issues}
}

}
\author{
\strong{Maintainer}: Damien Beillouin \email{d.beillouin@cirad.fr} (\href{https://orcid.org/{{orcid}}}{ORCID}) [copyright holder]

}
\keyword{internal}
